Errors out when no -P option was given
You committed a previous change of mine which checks whether "pigz" variable is non-zero. It turns out that the conditional expression now errors out on the second conditional now. There may be a more elegant way to handle, or I may not understand your intent, but here I check the exist status of the "which" command to determine whether to continue. I also changed the mbuffer line to make uniform.